Clarify FPDF_SaveAsCopy flags parameter
The documentation for the `flags` parameter in the `FPDF_SaveAsCopy` function in `fpdf_save.h` has been updated to clarify that 0 should be passed if no flags are applicable.

Additionally, in `mainJNILib.cpp`, the type of `widthInt` and `heightInt` variables has been changed from `jint` to `int` for local calculations before being used to create a Java `Size` object.